HSEARCH-3554 Adapt IndexSearchQueryIT to handle a higher number of documents
