+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/IndexSearchQueryBaseIT.java
@@ -1,0 +1,83 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.integrationtest.backend.tck.search;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.hibernate.search.engine.backend.document.IndexFieldReference;
+import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement;
+import org.hibernate.search.engine.backend.types.Sortable;
+import org.hibernate.search.engine.search.DocumentReference;
+import org.hibernate.search.engine.search.query.spi.IndexSearchQuery;
+import org.hibernate.search.integrationtest.backend.tck.testsupport.util.rule.SearchSetupHelper;
+import org.hibernate.search.util.impl.integrationtest.common.stub.mapper.StubMappingIndexManager;
+import org.hibernate.search.util.impl.integrationtest.common.stub.mapper.StubMappingSearchScope;
+
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+
+public class IndexSearchQueryBaseIT {
+
+	private static final String INDEX_NAME = "IndexName";
+
+	@Rule
+	public SearchSetupHelper setupHelper = new SearchSetupHelper();
+
+	private StubMappingIndexManager indexManager;
+
+	@Before
+	public void setup() {
+		setupHelper.withDefaultConfiguration()
+				.withIndex(
+						INDEX_NAME,
+						ctx -> new IndexMapping( ctx.getSchemaElement() ),
+						indexManager -> this.indexManager = indexManager
+				)
+				.setup();
+	}
+
+	@Test
+	public void getQueryString() {
+		StubMappingSearchScope scope = indexManager.createSearchScope();
+
+		IndexSearchQuery<DocumentReference> query = scope.query()
+				.asReference()
+				.predicate( f -> f.match().onField( "string" ).matching( "platypus" ) )
+				.toQuery();
+
+		assertThat( query.getQueryString() ).contains( "platypus" );
+	}
+
+	@Test
+	public void queryWrapper() {
+		StubMappingSearchScope scope = indexManager.createSearchScope();
+
+		QueryWrapper queryWrapper = scope.query()
+				.asReference( QueryWrapper::new )
+				.predicate( f -> f.match().onField( "string" ).matching( "platypus" ) )
+				.toQuery();
+		assertThat( queryWrapper.query.getQueryString() ).contains( "platypus" );
+	}
+
+	private static class IndexMapping {
+		final IndexFieldReference<String> string;
+
+		IndexMapping(IndexSchemaElement root) {
+			string = root.field( "string", f -> f.asString().sortable( Sortable.YES ) )
+					.toReference();
+		}
+	}
+
+	private static class QueryWrapper {
+		private final IndexSearchQuery<?> query;
+
+		private QueryWrapper(IndexSearchQuery<?> query) {
+			this.query = query;
+		}
+	}
+}
